requirements.txt
Babel==1.3
Faker==0.0.4
Flask==0.10.1
Flask-AppBuilder==1.2.1
Flask-Assets==0.10
Flask-BabelPkg==0.9.6
Flask-Bootstrap==3.3.0.1
Flask-Login==0.2.11
Flask-Migrate==1.3.0
Flask-Moment==0.4.0
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.10.3
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==0.9.8
SQLAlchemy-Searchable==0.8.0
SQLAlchemy-Utils==0.29.2
WTForms==2.0.1
Werkzeug==0.9.6
akismet==0.2.0
alembic==0.7.3
aniso8601==0.92
argparse==1.2.1
bleach==1.4.1
blinker==1.3
decorator==3.4.0
dirtools==0.2.0
fake-factory==0.4.2
-e git+https://github.com/saltycrane/flask-jquery-ajax-example.git@b99e670ce48bbb8b4d11372c17f9215131cd327e#egg=flask_jquery_ajax_example-master
globster==0.1.0
html5lib==0.999
itsdangerous==0.24
loremipsum==1.0.5
markdown2==2.3.0
psycopg2==2.5.4
pyparsing==2.0.3
python-openid==2.2.5
pytz==2014.10
requests==2.5.1
six==1.9.0
speaklater==1.3
validators==0.7
webassets==0.10.1
wsgiref==0.1.2
